Apply low retry period for the sms high priority lane on celery retry send
